adds an idea for ABS-Printed parts.
This includes a new techique to fix the magnets (.6mm at the closed front and a glueable Magnet fixer inside).
The LED-Case still has to be put onto the bottom plate as before. Though the case now looks fancier from outside.
And hopefully the front plate will fit better and hold itself (first tests support this).